# Nil out the slices to reduce alloc